+++ packages/adhere-ui-messagedialog/lib/types.d.ts
@@ -24,6 +24,6 @@ export interface IConfirmArgv extends IAlertArgv {
  */
 export interface IModalDialogProps {
     cloneBtn: boolean;
-    parent: HTMLElement | null;
     config: Object;
+    close(): void;
 }